third_party/blink/web_tests/fast/scroll-behavior/overflow-scroll-root-frame-animates-expected.txt

Tests that overflow scrolls on the root frame are animated.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.scrollY is not document.body.scrollHeight - window.innerHeight
PASS window.scrollY is > 0
PASS window.scrollY is > prevScrollY
PASS successfullyParsed is true

TEST COMPLETE
